Add a threshold for creating eageragg alternative
The eageragg alternative can sometimes end up producing a bad plan when the
ratio of the number of rows produced by the aggregate is high compared to the
number of rows produced by the join below it. This commit adds a threshold
above which the eageragg alternative will not be produced.