Fix Edm method for CPhysicalLimit
When we have a Global Limit operator which requests singleton distribution from
its child, it should have an `Exact` match of the distribution type from its
child. Otherwise we can end up with non-deterministic results. This might be
problematic when we are trying to insert into a replicated table for example.
This patch fixes the Edm method to do that.